+++ Runtime/ElementSelection/ISelectableElementVisualModule.cs
@@ -8,5 +8,7 @@ namespace UI.ElementSelection
 
         public void OnInitialized() {}
         public void OnSelectionChanged(bool isSelected);
+        public void OnHoverChanged(bool isHovered) {}
+        public float GetAnimationDuration() => 0f;
     }
 }