+++ Runtime/ElementSelection/VisualModules/GameObjectToggleSelectionVisualModule.cs
@@ -7,10 +7,35 @@ namespace UI.ElementSelection.VisualModules
     {
         [SerializeField] private GameObject _gameObject;
         [SerializeField] private bool _activateOnSelection = true;
+        [SerializeField] private SelectionModuleState<bool> _hoveredValue;
+
+        private bool _isSelected;
+        private bool _isHovered;
         
         public void OnSelectionChanged(bool isSelected)
         {
-            _gameObject.SetActive(_activateOnSelection ? isSelected : isSelected == false);
+            _isSelected = isSelected;
+            UpdateActive();
+        }
+
+        public void OnHoverChanged(bool isHovered)
+        {
+            _isHovered = isHovered;
+            UpdateActive();
+        }
+
+        private void UpdateActive()
+        {
+            if (_gameObject == null) return;
+
+            bool targetActive = _activateOnSelection ? _isSelected : !_isSelected;
+            
+            if (_isHovered && _hoveredValue.Enabled)
+            {
+                targetActive = _hoveredValue.Value;
+            }
+
+            _gameObject.SetActive(targetActive);
         }
     }
 }